refactor: FORMS-965 rename parameter validation code
