+++ app/tests/unit/forms/common/middleware/validateParameter.spec.js
@@ -1,0 +1,247 @@
+const { getMockReq, getMockRes } = require('@jest-mock/express');
+const { v4: uuidv4 } = require('uuid');
+
+const validateParameter = require('../../../../../src/forms/common/middleware/validateParameter');
+const formService = require('../../../../../src/forms/form/service');
+
+const formId = uuidv4();
+
+// Various types of invalid UUIDs that we see in API calls.
+const invalidUuids = [[''], ['undefined'], ['{{id}}'], ['${id}'], [uuidv4() + '.'], [' ' + uuidv4() + ' ']];
+
+afterEach(() => {
+  jest.clearAllMocks();
+});
+
+describe('validateFormId', () => {
+  describe('400 response when', () => {
+    test('formId is missing', async () => {
+      const req = getMockReq({
+        params: {},
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormId(req, res, next);
+
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+
+    test.each(invalidUuids)('formId is "%s"', async (eachFormId) => {
+      const req = getMockReq({
+        params: { formId: eachFormId },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormId(req, res, next, eachFormId);
+
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+  });
+
+  describe('allows', () => {
+    test('uuid for formId', async () => {
+      const req = getMockReq({
+        params: {
+          formId: formId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormId(req, res, next, formId);
+
+      expect(next).toHaveBeenCalledWith();
+    });
+  });
+});
+
+describe('validateFormVersionDraftId', () => {
+  const formVersionDraftId = uuidv4();
+
+  const mockReadDraftResponse = {
+    formId: formId,
+    id: formVersionDraftId,
+  };
+
+  formService.readDraft = jest.fn().mockReturnValue(mockReadDraftResponse);
+
+  describe('400 response when', () => {
+    test('formVersionDraftId is missing', async () => {
+      const req = getMockReq({
+        params: {
+          formId: formId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionDraftId(req, res, next);
+
+      expect(formService.readDraft).not.toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+
+    test.each(invalidUuids)('formVersionDraftId is "%s"', async (eachFormVersionDraftId) => {
+      const req = getMockReq({
+        params: { formId: formId, formVersionDraftId: eachFormVersionDraftId },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionDraftId(req, res, next, eachFormVersionDraftId);
+
+      expect(formService.readDraft).not.toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+  });
+
+  describe('404 response when', () => {
+    test('formId does not match', async () => {
+      formService.readDraft.mockReturnValueOnce({
+        formId: uuidv4(),
+        id: formVersionDraftId,
+      });
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionDraftId: formVersionDraftId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionDraftId(req, res, next, formVersionDraftId);
+
+      expect(formService.readDraft).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 404 }));
+    });
+  });
+
+  describe('handles error thrown by', () => {
+    test('readDraft', async () => {
+      const error = new Error();
+      formService.readDraft.mockRejectedValueOnce(error);
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionDraftId: formVersionDraftId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionDraftId(req, res, next, formVersionDraftId);
+
+      expect(formService.readDraft).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(error);
+    });
+  });
+
+  describe('allows', () => {
+    test('form version draft with matching form id', async () => {
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionDraftId: formVersionDraftId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionDraftId(req, res, next, formVersionDraftId);
+
+      expect(formService.readDraft).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith();
+    });
+  });
+});
+
+describe('validateFormVersionId', () => {
+  const formVersionId = uuidv4();
+
+  const mockReadVersionResponse = {
+    formId: formId,
+    id: formVersionId,
+  };
+
+  formService.readVersion = jest.fn().mockReturnValue(mockReadVersionResponse);
+
+  describe('400 response when', () => {
+    test('formVersionId is missing', async () => {
+      const req = getMockReq({
+        params: {
+          formId: formId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionId(req, res, next);
+
+      expect(formService.readVersion).not.toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+
+    test.each(invalidUuids)('formVersionId is "%s"', async (eachFormVersionId) => {
+      const req = getMockReq({
+        params: { formId: formId, formVersionId: eachFormVersionId },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionId(req, res, next, eachFormVersionId);
+
+      expect(formService.readVersion).not.toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 400 }));
+    });
+  });
+
+  describe('404 response when', () => {
+    test('formId does not match', async () => {
+      formService.readVersion.mockReturnValueOnce({
+        formId: uuidv4(),
+        id: formVersionId,
+      });
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionId: formVersionId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionId(req, res, next, formVersionId);
+
+      expect(formService.readVersion).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(expect.objectContaining({ status: 404 }));
+    });
+  });
+
+  describe('handles error thrown by', () => {
+    test('readVersion', async () => {
+      const error = new Error();
+      formService.readVersion.mockRejectedValueOnce(error);
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionId: formVersionId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionId(req, res, next, formVersionId);
+
+      expect(formService.readVersion).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith(error);
+    });
+  });
+
+  describe('allows', () => {
+    test('form version with matching form id', async () => {
+      const req = getMockReq({
+        params: {
+          formId: formId,
+          formVersionId: formVersionId,
+        },
+      });
+      const { res, next } = getMockRes();
+
+      await validateParameter.validateFormVersionId(req, res, next, formVersionId);
+
+      expect(formService.readVersion).toHaveBeenCalled();
+      expect(next).toHaveBeenCalledWith();
+    });
+  });
+});
